Add files via upload
To run the main.py without editing the files paths: Under a folder called flash_cards(or anything you want), place your main.py in there. Then create a folder called "data" and another called "images" in the flash_cards folder. In images, place the 4 .png files and in data place the .csv so that the main.py runs correctly